pyproject.toml
[project]
name = "fgp-bot-discord"
version = "0.1.0"
description = "Discord bot 'file manager' for personal uses."
authors = [
    { name = "ArtiArtem", email = "ArtiArtem8@users.noreply.github.com" },
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aiolimiter>=1.2.1",
    "aiosqlite>=0.21.0",
    "discord-py>=2.5.2",
    "dotenv>=0.9.9",
    "fuzzywuzzy>=0.18.0",
    "pillow>=11.2.1",
    "pydantic>=2.11.4",
]

[tool.pytest.ini_options]
pythonpath = "."
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = ["-v", "--tb=short", "--strict-markers", "-ra"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "benchmark: marks tests as performance benchmarks",
    "integration: marks tests as integration tests",
]
filterwarnings = ["ignore::DeprecationWarning:discord.player"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-benchmark>=5.1.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.11.9",
]

[tool.ruff]
exclude = [".git", ".ruff_cache", ".venv", ".vscode", "site-packages", "tests"]
line-length = 88
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = ["ALL"]
fixable = ["ALL"]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs,tools}/*" = ["E402"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"
